requirements.txt
requests-cache[all]
pytest
ccxt
click